pyproject.toml
[project]
name = "mcp-gitlab"
version = "0.1.0"
description = "MCP server for GitLab integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "python-gitlab>=4.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-gitlab = "mcp_gitlab.__main__:main"